refactor(app): centralize preview writeability policy
