+++ src/app/policy/write/write_guardrails.rs
@@ -1,5 +1,104 @@
-use crate::domain::QueryValue;
+use crate::domain::{DatabaseType, QueryResult, QueryValue, Table, TableKind};
 use crate::policy::sql::statement_classifier::StatementKind;
+use crate::policy::write::write_update::build_pk_pairs;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum StableRowIdentity {
+    PrimaryKey(Vec<String>),
+    SqliteRowid { alias: String },
+}
+
+impl StableRowIdentity {
+    pub fn identity_pairs_for_row(
+        &self,
+        result: &QueryResult,
+        row_idx: usize,
+    ) -> Option<Vec<(String, QueryValue)>> {
+        match self {
+            Self::PrimaryKey(columns) => {
+                let row = result.values().get(row_idx)?;
+                build_pk_pairs(&result.columns, row, columns)
+            }
+            Self::SqliteRowid { alias } => result
+                .hidden_value_at(row_idx, alias)
+                .cloned()
+                .map(|value| vec![(alias.clone(), value)]),
+        }
+    }
+
+    pub fn predicate_pairs_for_row(
+        &self,
+        result: &QueryResult,
+        row_idx: usize,
+    ) -> Option<Vec<(String, QueryValue)>> {
+        match self {
+            Self::PrimaryKey(_) => self.identity_pairs_for_row(result, row_idx),
+            Self::SqliteRowid { alias } => {
+                let rowid = result.hidden_value_at(row_idx, alias)?.clone();
+                let row = result.values().get(row_idx)?;
+                if row.is_empty() || result.columns.is_empty() || row.len() != result.columns.len()
+                {
+                    return None;
+                }
+                let mut pairs = Vec::with_capacity(row.len() + 1);
+                pairs.push((alias.clone(), rowid));
+                pairs.extend(result.columns.iter().cloned().zip(row.iter().cloned()));
+                Some(pairs)
+            }
+        }
+    }
+
+    pub fn is_primary_key_column(&self, column_name: &str) -> bool {
+        match self {
+            Self::PrimaryKey(columns) => columns.iter().any(|pk| pk == column_name),
+            Self::SqliteRowid { alias: _ } => false,
+        }
+    }
+
+    pub fn uses_sqlite_rowid(&self) -> bool {
+        matches!(self, Self::SqliteRowid { .. })
+    }
+}
+
+pub fn stable_row_identity_for_table(
+    database_type: DatabaseType,
+    table: &Table,
+) -> Option<StableRowIdentity> {
+    if let Some(pk) = table.primary_key.as_ref().filter(|cols| !cols.is_empty()) {
+        return Some(StableRowIdentity::PrimaryKey(pk.clone()));
+    }
+    if database_type != DatabaseType::SQLite {
+        return None;
+    }
+    table
+        .sqlite_rowid_alias()
+        .map(|alias| StableRowIdentity::SqliteRowid {
+            alias: alias.to_string(),
+        })
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum PreviewWriteability {
+    Writable,
+    ReadOnly(&'static str),
+    MissingStableRowIdentity,
+}
+
+pub fn preview_writeability(database_type: DatabaseType, table: &Table) -> PreviewWriteability {
+    if table.kind_info.kind == TableKind::View {
+        return PreviewWriteability::ReadOnly("view");
+    }
+    if table.kind_info.kind == TableKind::Virtual {
+        return PreviewWriteability::ReadOnly("virtual table");
+    }
+    if table.kind_info.without_rowid {
+        return PreviewWriteability::ReadOnly("WITHOUT ROWID table");
+    }
+    if stable_row_identity_for_table(database_type, table).is_none() {
+        return PreviewWriteability::MissingStableRowIdentity;
+    }
+    PreviewWriteability::Writable
+}
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum WriteOperation {
